do folded_upcast in to_indexed_op, real_size uses views_to_indexed_ops
